:hammer: refactor-move test cases to single class
