+++ app/src/test/java/com/example/mobileiassistapp/MarkingActivityTest.java
@@ -5,11 +5,6 @@ import org.junit.Test;
 import static org.junit.Assert.*;
 
 public class MarkingActivityTest {
-
-    @Test
-    public void onCreate() {
-    }
-
     @Test
     public void calculateCAMarks() {
         float assg1 = 70;
@@ -24,4 +19,32 @@ public class MarkingActivityTest {
 
         assertEquals(expected,output,delta);
     }
+
+    @Test
+    public void calculateTotal() {
+        float ca = 50;
+        float final_marks= 44;
+        float expect = 94;
+        float delta = Float.parseFloat("0.1");
+
+        Marking2Activity marking2Activity = new Marking2Activity();
+
+        float output = marking2Activity.calculateTotal(final_marks,ca);
+
+        assertEquals(expect,output,delta);
+    }
+
+    @Test
+    public void calculateFinalMarks() {
+        float finalMarksOutOf100 = 70;
+        float expect = 35;
+        float delta = Float.parseFloat("0.1");
+
+        MarkingActivity markingActivity = new MarkingActivity();
+        float output = markingActivity.calculateFinalMarks(finalMarksOutOf100);
+
+        assertEquals(expect,output,delta);
+    }
+
+
 }
